Move ODE definition into subfolder
