+++ dymos/examples/robertson_problem/doc/test_doc_robertson_problem.py
@@ -1,4 +1,3 @@
-import os
 import unittest
 
 import matplotlib
@@ -7,7 +6,7 @@ import matplotlib.pyplot as plt
 plt.style.use('ggplot')
 
 from openmdao.utils.testing_utils import use_tempdirs
-from dymos.examples.robertson_problem.robertson_ode import RobertsonODE
+from dymos.examples.robertson_problem.doc.robertson_ode import RobertsonODE
 
 @use_tempdirs
 class TestRobertsonProblemForDocs(unittest.TestCase):
